Renamed `Store` to `VersionedStore`
Added skeleton ABC for `Store`
