+++ openghg/store/storage/__init__.py
@@ -1,5 +1,5 @@
 from ._chunking import ChunkingSchema
-from ._store import VersionedStore
+from .versioned_store import VersionedStore
 from ._localzarrstore import LocalZarrStore
 from ._encoding import get_zarr_encoding
 from ._compression import compare_compression
